lavisco/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-jazzmin==2.4.7
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-tinymce==3.3.0
djangorestframework==3.12.4
drf-registration==0.1.2
idna==2.10
Pillow==8.2.0
pkg-resources==0.0.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
